fix implicit conversion float to double (-Wdouble-promotion error) and old-style-cast error

+++ MMVII/include/uti_fonc_analytique.h
@@ -25,7 +25,7 @@ template <typename Type> Type DerASin(const Type & aSin)
 {
    Type UMS2 = 1-Square(aSin);
    MMVII_ASSERT_STRICT_POS_VALUE(UMS2);
-   return 1.0 / std::sqrt(UMS2);
+   return 1.0 / static_cast<double> (std::sqrt(UMS2));
 }
 
 template <typename Type> Type ASin(const Type & aSin)
@@ -56,7 +56,7 @@ template <typename Type> Type DerSinC(const Type & aTeta,const Type & aEps)
    Type aT9 = aT7   * aTeta2;
 
 
-   return  - aTeta*(2.0/Fact3) + aT3 *(4.0/Fact5) - aT5*(6.0/Fact7) + aT7*(8.0/Fact9) - aT9 * (10.0/Fact11);
+   return  - aTeta*(2.0/static_cast<double>(Fact3)) + aT3 *(4.0/static_cast<double>(Fact5)) - aT5*(6.0/static_cast<double>(Fact7)) + aT7*(8.0/static_cast<double>(Fact9)) - aT9 * (10.0/static_cast<double>(Fact11));
 }
 template <typename Type> Type DerSinC(const Type & aTeta)
 {
@@ -74,7 +74,7 @@ template <typename Type> Type sinC(const Type & aTeta,const Type & aEps)
    Type aT6 = aT4 * aT2;
    Type aT8 = Square(aT4);
 
-   return 1.0 - aT2/Fact3 + aT4/Fact5 - aT6/Fact7 + aT8/Fact9;
+   return 1.0 - aT2/static_cast<double>(Fact3) + aT4/static_cast<double>(Fact5) - aT6/static_cast<double>(Fact7) + aT8/static_cast<double>(Fact9);
 }
 template <typename Type> Type sinC(const Type & aTeta)
 {
@@ -92,7 +92,7 @@ template <typename Type> Type AtanXsY_sX(const Type & X,const Type & Y,const Typ
    Type XsY6 = XsY4 * XsY2;
    Type XsY8 = XsY4 * XsY4;
 
-   return (1 -XsY2/3.0 + XsY4/5.0 -XsY6/7.0 + XsY8/9.0) / Y;
+   return (1 -static_cast<double>(XsY2)/3.0 + static_cast<double>(XsY4)/5.0 -static_cast<double>(XsY6)/7.0 + static_cast<double>(XsY8)/9.0) / static_cast<double>(Y);
 }
 template <typename Type> Type AtanXsY_sX(const Type & X,const Type & Y)
 {
